Separate the xDripViewer From the WixelReader.
